Update swagger.md - Round #1
I re-worded A LOT here, and I also wasn't sure if "resource" in "API resource" needs to be capitalized or not, so I left it alone. Lastly, I notice this chapter is called "swagger" but I think I only see that term pop up in the very beginning and the very end. Not sure if that term should be a touchpoint somewhere else towards the middle or not.