Update subresources.md - Round #1
This one was LONG and, at times, hard to follow. I did my best to make better sense of some of the confusing moments, and I definitely think this needs a different ending, since I know there's another chapter after this. The transition there could be better to set up the next chapter, but I'm not exactly sure what to say.