Update login-response.md - Round #1
I wasn't sure if "user" needed capitalization or not. Also, I'm not sure if you've noticed this in any scripts you've edited yet, but there was a moment or two throughout the script (mostly towards the end, I think) where a code snippet was thrown in randomly, but it didn't look like it belonged and you didn't address it verbally. Was this just some weird hiccup with Temi? I went ahead and removed it.